core/standards/css3/image-fit/js/ua-stylesheet/video/004.html

<!doctype html>
<title>video default style, src</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<video src=../../support/240x160.ogv></video>
<script>
var t = async_test('load', {timeout:2000});
on_event(window, 'load', function() { t.step(function() {
  var e = document.body.firstChild;
  assert_equals(getComputedStyle(e, '').OObjectFit, 'contain');
  assert_equals(getComputedStyle(e, '').OObjectPosition, '50% 50%');
  t.done();
}); });
</script>

